Converted diff to a double before calculating time difference in hours

+++ src/main/events/MailDelivery.java
@@ -123,8 +123,8 @@ public class MailDelivery extends BusinessEvent {
 
 		// Get msec from each, and subtract.
 		long diff = d2.getTime() - d1.getTime();
-		long diffMinutes = (long) (diff / (60 * 1000)) % 60;
-		double diffHours = (diff / (60 * 60 * 1000));
+		long diffMinutes = (long)(diff / (60 * 1000)) % 60;
+		double diffHours = ((double)diff / (60 * 60 * 1000));
 		long diffDays = (long) (diff / (24 * 60 * 60 * 1000));
 		System.out.println("Time in minutes: " + diffMinutes + " minutes.");
 		System.out.println("Time in hours: " + diffHours%24 + " hours.");
